% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_performance.R
\name{player_performance}
\alias{player_performance}
\title{Player Performance Score}
\usage{
player_performance(goals, assists, passes, tackles, shots)
}
\arguments{
\item{goals}{Number of goals scored.}

\item{assists}{Number of assists.}

\item{passes}{Total completed passes.}

\item{tackles}{Number of tackles.}

\item{shots}{Total shots taken.}
}
\value{
A numeric performance score.
}
\description{
This function calculates a player's performance score based on goals, assists, passes, tackles, and shots.
}
\examples{
player_performance(goals = 3, assists = 2, passes = 50, tackles = 4, shots = 6)
}
